feat: Add RaceResult
